==> 04_prophage_rumen_mags/01_check_prophage.ipynb <==
import os
import pandas as pd
import glob
import re
from Bio import SeqIO

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/04_prophage_rumen_mags")

# category1: all viral contigs identified from MAGs 
category1 = []
records = SeqIO.parse("prophage_rumen_mags_vs2_vibrant_intersection.fa", "fasta")
for record in records:
    category1.append(record.name)
    
    


prophage_checkv = []
checkv_results = glob.glob("**/proviruses.fna", recursive = True)
for result in checkv_results:
    records = SeqIO.parse(result, "fasta")
    for record in records:
        prophage_checkv.append(record.name)

vibrant_out = pd.read_csv("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/04_prophage_rumen_mags/vibrant_out/VIBRANT_prophage_rumen_mags_vs2_keep1_2/VIBRANT_results_prophage_rumen_mags_vs2_keep1_2/VIBRANT_genome_quality_prophage_rumen_mags_vs2_keep1_2.tsv", sep = "\t")

vibrant_df = pd.read_csv('vibrant_out/VIBRANT_prophage_rumen_mags_vs2_keep1_2/VIBRANT_results_prophage_rumen_mags_vs2_keep1_2/VIBRANT_genbank_table_prophage_rumen_mags_vs2_keep1_2.tsv', sep ='\t')

def find_marker(f):
    if re.search(r'integrase|excisionase|repressor protein', f.lower()):
        return True
    else:
        return False
    

marker_match = set([f.split(' ')[0] for f in vibrant_df[vibrant_df.name.apply(find_marker)]['scaffold'] if f in category1])

# category2: excised from host contigs or contain prophage marker, for viral contigs in category1
category2 = list(set(marker_match | set(prophage_checkv)) & set(category1))

# category3: excised from host contigs 
category3 = list(set(prophage_checkv) & set(category1))

# category4: excised from host contigs and contain prophage marker
#category4 = list(set(prophage_checkv) & marker_match)

with open("prophage_rumen_mags_category1.txt", 'w') as file:
    for f in category1:
        file.write(f + "\n")

with open("prophage_rumen_mags_category2.txt", 'w') as file:
    for f in category2:
        file.write(f + "\n")
        
with open("prophage_rumen_mags_category3.txt", 'w') as file:
    for f in category3:
        file.write(f + "\n")   
        
#with open("prophage_rumen_mags_category4.txt", 'w') as file:
#    for f in category4:
#        file.write(f + "\n")
