==> 04_prophage_rumen_mags/03_prepare_for_Propagate.ipynb <==
## Customize file required for Propagate



import os
import glob
import pandas as pd
import re
from Bio import SeqIO
from Bio.Seq import Seq
from Bio.SeqRecord import SeqRecord

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/04_prophage_rumen_mags")

df = pd.read_csv("rumen_mags_prophage_summary.csv")

*Propagate input fasta definition line cannot have pipe symbol so modifiy original Vibrant output*

category3 = []
with open("prophage_rumen_mags_category3.txt", 'r') as file:
    lines = file.readlines()
    for f in lines:
        category3.append(f.strip())

checkv_prophage = glob.glob("checkv/**/proviruses.fna", recursive = True)

prophage_checkv = {}
for f in checkv_prophage:
    records = SeqIO.parse(f, "fasta")
    for record in records:
        range = {}
        if record.name in category3:
            prophage = record.name
            range["start"] = record.description.split(' ')[1].split('/')[0].split('-')[0]
            range["end"] = record.description.split(' ')[1].split('/')[0].split('-')[1]
            prophage_checkv[prophage] = range

propagate_df = pd.DataFrame.from_dict(prophage_checkv, orient = "index").reset_index()

for index, row in propagate_df.iterrows():
    prophage = row["index"]
    scaffold = prophage.split('|')[0]
    propagate_df.loc[index, "scaffold"] = scaffold

propagate_df.rename(columns = {"index":"fragment", "end":"stop"}, inplace = True)

propagate_df = propagate_df[["scaffold", "fragment", "start", "stop"]]

host_genome = pd.read_csv("rumen_mags_prophage_summary.csv")

host_df = pd.merge(propagate_df, host_genome, left_on = "fragment", right_on = "prophage", how = "left").drop("prophage", axis = 1)

host_genomes = ['../../../databases/rumen_mags_high_quality/dereplicated_genomes_with_taxid/' + f + '.fasta' for f in host_df.host_genome]

with open("rumen_mags_prophage_scaffolds_for_propogate.fasta", 'w') as outfile:
    for genome in host_genomes:
        _ = genome.split('/')[-1].split('.fasta')[0]
        records = SeqIO.parse(genome, "fasta")
        seq = list(host_df.query('host_genome == @_')['scaffold'])[0]     
        for record in records:
            if record.description.split(" ")[1] == seq:
                newrecord = SeqRecord(
                Seq(record.seq),
                id=seq,
                description = "")
                
                SeqIO.write(newrecord, outfile, "fasta")
                break
    

propagate_df.to_csv("prophage_scaffold_coordinates_for_propogate.tsv", sep = "\t", index = None)
