==> 04_prophage_rumen_mags/05_non_cryptic_prophage_propogate.ipynb <==
# Non-cryptic if found active in any sample by Propogate

import os
import glob
import pandas as pd
import re
from Bio import SeqIO

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/04_prophage_rumen_mags/propogate_out")

propogate_out = glob.glob("**/*.tsv")

active_phage = []
for f in propogate_out:
    df = pd.read_csv(f"{f}", sep = '\t')
    active_phage = active_phage + list(df.query('active == "active"')['prophage'])
   

with open("../non-cryptic-prophage.txt", 'w') as outfile:
    for f in set(active_phage):
        outfile.write(f + "\n")

with open("../non-cryptic-prophage.fasta", 'w') as outfile:
    records = SeqIO.parse("../prophage_rumen_mags_category2.fasta", "fasta")
    for record in records:
        if record.name in set(active_phage):
            SeqIO.write(record, outfile, "fasta")
