==> 04_prophage_rumen_mags/06_prophage_taxa.ipynb <==
import os
import pandas as pd
import re
from Bio import SeqIO

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/04_prophage_rumen_mags/")

taxa = pd.read_csv("prophage_votus_taxa_phagcn_old_ictv.csv")[["contig_name", "prediction"]]

taxa_df = taxa.groupby('prediction').prediction.count().reset_index(name = "n")

votus = []
records = SeqIO.parse("prophage_rumen_mags_vOTUs.fa", "fasta")
for record in records:
    votus.append(record.id)
n_votus = len(votus)    

taxa_df.loc[6, "prediction"] = "Unassigned"
taxa_df.loc[6, "n"] = n_votus - len(taxa)


taxa_df

with open("non-cryptic-prophage.txt", 'r') as file:
    non_cryptic = file.read().split("\n")

n_non_cryptic = len(non_cryptic)

host = pd.read_csv("rumen_mags_prophage_summary.csv")


n_genome = len(set(host.user_genome))

n_species =len(set(host.Species))

with open("prophage_rumen_mags_category2.txt", 'r') as file:
    category3 = file.read().split("\n")
    
n_prophage = len(category3)

n_species_with_prophage = len(set(host.query('prophage in @category3').Species))

n_genome_with_prophage = len(set(host.query('prophage in @category3').user_genome))

print(f'{n_prophage} prophage ({n_non_cryptic} non-cryptic) out of {n_genome} host genomes')
print(f'{n_species_with_prophage} host species with prophage out of {n_species} host species')

prophage_prevalence = host.groupby("user_genome").prophage.count().reset_index().groupby('prophage').prophage.count().reset_index(name = "prevelance").rename(columns = {"prophage": "prophage_per_genome"})

prophage_prevalence

prophage_prevalence.to_csv("prophage_prevalence.csv", index = None)
taxa_df.to_csv("prophage_taxa_df.csv", index = None)
