==> 04_prophage_rumen_mags/07_itol_tree_annotation.ipynb <==
import os
import pandas as pd
import re
import seaborn as sns

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/04_prophage_rumen_mags/")

host = pd.read_csv("rumen_mags_prophage_summary.csv")

unclassified = list(host.query('Species != Species').index)
for index, row in host.iterrows():
    if index in unclassified:
        host.loc[index, "Species"] = "Unclassfied " + str(host.loc[index, "Genera"])

with open("prophage_rumen_mags_category3.txt", 'r') as file:
    category3 = file.read().split("\n")

core_genus = set(pd.read_csv("../R_project/core_genus_taxa.csv")["Genera"])
core_species = set(pd.read_csv("../R_project/core_species_taxa.csv")["Species"])
archaea_phylum = ["Methanobacteriota", "Thermoplasmatota"]

n_host_species_bac = len(set(host.query('Genera in @core_genus').query('Phylum not in @archaea_phylum').Species))
n_host_genome_bac = len(set(host.query('Genera in @core_genus').query('Phylum not in @archaea_phylum').user_genome))
n_host_genome_bac_with_prophage = len(set(host.query('Genera in @core_genus').query('Phylum not in @archaea_phylum').query('prophage in @category3').user_genome))
n_host_species_bac_with_prophage = len(set(host.query('Genera in @core_genus').query('Phylum not in @archaea_phylum').query('prophage in @category3').Species))

host = host.query('Genera in @core_genus').query('Phylum not in @archaea_phylum')

n_host_genome_bac

n_host_species_bac

n_host_species_bac_with_prophage

n_host_genome_bac_with_prophage

n_prophage_per_species = host.groupby("Species").prophage.count().reset_index()

with open("non-cryptic-prophage.txt", 'r') as file:
    active_prophage = file.read().split("\n")

def _(x):
    if x in active_prophage:
        return(True)
    else:
        return(False)

host.loc[host[list(host.prophage.apply(_))].index, "non_cryptic"] = "True"

bac_for_tree = set(host.query('prophage in @category3').query('Phylum not in @archaea_phylum').query('Genera in @core_genus').Species)

bac_genome_for_tree = []
for f in bac_for_tree:
    genome =  list(host.query('Species == @f')['user_genome'])[0] 
    bac_genome_for_tree.append(genome)

bac_genome_for_tree_df =  host.query('user_genome in @bac_genome_for_tree')[['Species', 'user_genome', "Phylum"]].drop_duplicates()

n_non_cryptic_per_species = host.groupby("Species").non_cryptic.count().reset_index()

annotate_pre = pd.merge(bac_genome_for_tree_df, n_non_cryptic_per_species, left_on = "Species", right_on = "Species")

tmp1 = host.query('Species in @bac_for_tree').groupby("Species").prophage.count().reset_index().rename(columns = {"prophage":"n_prophage"})

tmp2 = host.query('Species in @bac_for_tree').groupby("Species").user_genome.count().reset_index().rename(columns = {"user_genome":"n_user_genome"})

prophage_rate = pd.merge(tmp1, tmp2, left_on = "Species", right_on = "Species")

prophage_rate["prophage_rate"] = prophage_rate.n_prophage/prophage_rate.n_user_genome

prophage_rate_df = prophage_rate.sort_values("prophage_rate", ascending=False)

annotate_df = pd.merge(annotate_pre, prophage_rate_df, left_on = "Species", right_on = "Species" )

annotate_df["non_cryptic_rate"] = annotate_df.non_cryptic/annotate_df.n_prophage

annotate_df

len(set(annotate_df.Phylum))

print(sns.color_palette("tab10").as_hex())

palette = ['#1f77b4', '#ff7f0e', '#2ca02c', '#d62728', '#9467bd', '#8c564b', '#e377c2', '#7f7f7f', '#bcbd22', '#17becf'][0:9]

phylum_color = {}
for f in range(len(set(annotate_df.Phylum))):
    phylum_color[list(set(annotate_df.Phylum))[f]] = palette[f]
    

phylum_color

for index, row in annotate_df.iterrows():
    phylum = row["Phylum"]
    Species = row["Species"]
    color = phylum_color[phylum]
    annotate_df.loc[index, "Phylum_color"] = color
    if Species in core_species:
        annotate_df.loc[index, "core_species"] = True
    else:
        annotate_df.loc[index, "core_species"] = False
        

annotate_df

with open("itol_host_phylum_color_strip.txt", "w") as outfile:
    with open("itol_host_phylum_color_strip_template.txt", "r") as infile:
        outfile.write(infile.read())
        outfile.write('\n')
        for index, row in annotate_df.iterrows():
            user_genome = row['user_genome']
            phylum = row['Phylum_color']
            outfile.write(user_genome + ' ' + phylum + "\n")

with open("itol_prophage_per_genome.txt", "w") as outfile:
    with open("itol_prophage_gradient_template.txt", "r") as infile:
        outfile.write(infile.read())
        outfile.write('\n')
        for index, row in annotate_df.iterrows():
            user_genome = row['user_genome']
            prophage_rate = row['prophage_rate']
            outfile.write(user_genome + ' ' + str(prophage_rate) + "\n")

with open("itol_non_cryptic_per_prophage.txt", "w") as outfile:
    with open("itol_prophage_gradient_template.txt", "r") as infile:
        outfile.write(infile.read())
        outfile.write('\n')
        for index, row in annotate_df.iterrows():
            user_genome = row['user_genome']
            non_cryptic_rate = row['non_cryptic_rate']
            outfile.write(user_genome + ' ' + str(non_cryptic_rate) + "\n")

with open("itol_prophage_core_species.txt", "w") as outfile:
    with open("itol_prophage_core_species_template.txt", "r") as infile:
        outfile.write(infile.read())
        outfile.write('\n')
        for index, row in annotate_df.iterrows():
            user_genome = row['user_genome']
            core_species = row['core_species']
            if core_species:
                outfile.write(user_genome + ',' + str(1) + "\n")
