==> 04_prophage_rumen_mags/08_prophage_amr.ipynb <==
import os
import pandas as pd
import re

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/04_prophage_rumen_mags/amr/")

card = pd.read_csv("prophage_rumen_mags_category2_card.bln", sep = "\t", names = 'qseqid sseqid pident length qlen slen mismatch gapopen qstart qend sstart send evalue bitscore'.split(' '))

hits = card.query('(qend - qstart)/slen > 0.4').query('pident > 80')

contigs_hits =[re.search(r"(.*)_[0-9]*", f).group(1) for f in list(set(hits.qseqid))]

with open('amr_containing_contigs.txt', 'w') as outfile:
    for f in contigs_hits:
        outfile.write(f + '\n')
