==> 04_prophage_rumen_mags/10_filtered_amr_manual_check.ipynb <==
import os
import pandas as pd
import re

os.chdir("/fs/scratch/PAS0439/Ming/virome_ecology_core_prkaryotes/results/04_prophage_rumen_mags/")

with open("prophage_amr_contigs_final.txt", 'r') as file:
    amr = file.read().split("\n")

amr_out = pd.read_csv("amr/prophage_rumen_mags_category2_card.bln", sep = '\t', names = 'qseqid sseqid pident length qlen slen mismatch gapopen qstart qend sstart send evalue bitscore'.split(' '))

def seperate_id(x):
    out = re.search(r'(.*)_([0-9])+', x)
    contig = out.group(1)
    gene_id = out.group(2)
    return contig, gene_id


for index, row in amr_out.iterrows():
    amr_out.loc[index, ["contig", "gene_id"]] = seperate_id(row["qseqid"])

# Due to multiple hits (similar amr), only choose amr with the highest sequence identity as the true hit

amr_out.loc[amr_out.query('contig in @amr').groupby(['contig',  'gene_id']).pident.idxmax()]

genomic_context = pd.read_csv("prophage_amr_for_manual_curation.tsv", sep = "\t")

pd.set_option('display.max_rows', None)

genomic_context.query('fasta in @amr').to_csv('amr_genomic_context_manual_check.tsv', sep = "\t", index = None)
